+++ configuration/interface.go
@@ -1,5 +1,6 @@
 //
 // Copyright (c) 2019 Intel Corporation
+// Copyright (C) 2023 IOTech Ltd
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,9 +17,7 @@
 
 package configuration
 
-import (
-	"github.com/pelletier/go-toml"
-)
+import "github.com/edgexfoundry/go-mod-messaging/v3/messaging"
 
 type Client interface {
 	// HasConfiguration checks to see if the Configuration service contains the service's configuration.
@@ -27,8 +26,9 @@ type Client interface {
 	// HasSubConfiguration checks to see if the Configuration service contains the service's sub configuration.
 	HasSubConfiguration(name string) (bool, error)
 
-	// PutConfigurationToml puts a full toml configuration into the Configuration service
-	PutConfigurationToml(configuration *toml.Tree, overwrite bool) error
+	// PutConfigurationMap puts a full map configuration into the Configuration service
+	// The sub-paths to where the values are to be stored in the Configuration service are generated from the map key.
+	PutConfigurationMap(configuration map[string]any, overwrite bool) error
 
 	// PutConfiguration puts a full configuration struct into the Configuration service
 	PutConfiguration(configStruct interface{}, overwrite bool) error
@@ -41,7 +41,7 @@ type Client interface {
 	// WatchForChanges sets up a Consul watch for the target key and send back updates on the update channel.
 	// Passed in struct is only a reference for Configuration service, empty struct is ok
 	// Sends the configuration in the target struct as interface{} on updateChannel, which caller must cast
-	WatchForChanges(updateChannel chan<- interface{}, errorChannel chan<- error, configuration interface{}, waitKey string)
+	WatchForChanges(updateChannel chan<- interface{}, errorChannel chan<- error, configuration interface{}, waitKey string, msgClient messaging.MessageClient)
 
 	// StopWatching causes all WatchForChanges processing to stop and waits until they have stopped.
 	StopWatching()
@@ -55,6 +55,12 @@ type Client interface {
 	// GetConfigurationValue gets a specific configuration value from the Configuration service
 	GetConfigurationValue(name string) ([]byte, error)
 
+	// GetConfigurationValueByFullPath gets a specific configuration value from the Configuration service
+	GetConfigurationValueByFullPath(fullPath string) ([]byte, error)
+
 	// PutConfigurationValue puts a specific configuration value into the Configuration service
 	PutConfigurationValue(name string, value []byte) error
+
+	// GetConfigurationKeys returns all keys under name
+	GetConfigurationKeys(name string) ([]string, error)
 }