+++ internal/pkg/keeper/utils/http/common.go
@@ -10,7 +10,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net"
 	"net/http"
 	"net/url"
@@ -23,7 +23,7 @@ type ErrorResponse struct {
 
 // Helper method to get the body from the response after making the request
 func getBody(resp *http.Response) ([]byte, error) {
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return body, errors.New("failed to get the body from the response")
 	}
@@ -37,7 +37,7 @@ func makeRequest(req *http.Request) (*http.Response, error) {
 	if err != nil {
 		var netErr *net.OpError
 		if errors.As(err, &netErr) {
-			return nil, errors.New(fmt.Sprintf("%s cannot be reached, this service is not available.", req.URL.Host))
+			return nil, fmt.Errorf("%s cannot be reached, this service is not available", req.URL.Host)
 		} else {
 			return nil, errors.New("failed to send a http request")
 		}
@@ -66,12 +66,12 @@ func createRequest(httpMethod string, baseUrl string, requestPath string, reques
 
 // sendRequest will make a request with raw data to the specified URL.
 // It returns the body as a byte array if successful and an error otherwise.
-func sendRequest(req *http.Request) ([]byte, ErrorResponse) {
+func sendRequest(req *http.Request) ([]byte, ErrorResponse, error) {
 	var errResponse ErrorResponse
 
 	resp, err := makeRequest(req)
 	if err != nil {
-		return nil, errResponse
+		return nil, errResponse, err
 	}
 	defer func() {
 		_ = resp.Body.Close()
@@ -79,20 +79,20 @@ func sendRequest(req *http.Request) ([]byte, ErrorResponse) {
 
 	bodyBytes, err := getBody(resp)
 	if err != nil {
-		return nil, errResponse
+		return nil, errResponse, err
 	}
 
 	if resp.StatusCode <= http.StatusMultiStatus {
-		return bodyBytes, errResponse
+		return bodyBytes, errResponse, nil
 	}
 
 	// Handle error response
-	e := json.Unmarshal(bodyBytes, &errResponse)
-	if e != nil {
-		return nil, errResponse
+	err = json.Unmarshal(bodyBytes, &errResponse)
+	if err != nil {
+		return nil, errResponse, err
 	}
 
-	return nil, errResponse
+	return nil, errResponse, nil
 }
 
 func createRequestWithRawData(httpMethod string, baseUrl string, requestPath string, requestParams url.Values, data interface{}) (*http.Request, error) {